feat(1670): move function to util
